+++ tests/correctness/targets/FilteredCopy_I18N/Input/test.xpybuild.py
@@ -39,4 +39,9 @@ FilteredCopy('${OUTPUT_DIR}/copy-default.json', '${OUTPUT_DIR}/writefile-default
 FilteredCopy('${OUTPUT_DIR}/copy-customized.json', '${OUTPUT_DIR}/writefile-customized.foo.yaml', [StringReplaceLineMapper('Text', 'Replaced text'),]).option('common.fileEncodingDecider',
 	ExtensionBasedFileEncodingDecider({'.${Y}aml':'iso-8859-1','.json':'iso-8859-1'}, None))
 
-FilteredCopy('${OUTPUT_DIR}/copy-compressed.jpg', '${OUTPUT_DIR}/writefile-compressed.jpg') # rely on mime types to detect this as binary
+FilteredCopy('${OUTPUT_DIR}/copy-compressed.jpg', '${OUTPUT_DIR}/writefile-compressed.jpg' # rely on mime types in the default decider to detect this as binary
+	).option('common.fileEncodingDecider',
+		ExtensionBasedFileEncodingDecider(
+				{'.foo': 'utf-8', '.bar': ExtensionBasedFileEncodingDecider.BINARY}, 
+				default=ExtensionBasedFileEncodingDecider.getDefaultFileEncodingDecider())
+	)